+++ wp-admin/includes/plugin-install.php
@@ -14,19 +14,40 @@
  * Gets the markup for the plugin install action button.
  *
  * @param string       $name           Plugin name.
- * @param array|object $data           Plugin data.
- * @param bool         $compatible_php PHP compatibility check.
- * @param bool         $compatible_wp  WP compatibilty check.
+ * @param array|object $data           {
+ *     An array or object of plugin data. Can be retrieved from the API.
+ *
+ *     @type string   $slug             The plugin slug.
+ *     @type string[] $requires_plugins An array of plugin dependency slugs.
+ *     @type string   $version          The plugin's version string. Used when getting the install status.
+ * }
+ * @param bool         $compatible_php   The result of a PHP compatibility check.
+ * @param bool         $compatible_wp    The result of a WP compatibility check.
  *
  * @return string $button The markup for the dependency row button.
  */
 function wp_get_plugin_action_button( $name, $data, $compatible_php, $compatible_wp ) {
-	$button = '';
-	$status = install_plugin_install_status( $data );
+	$button                = '';
+	$data                  = (object) $data;
+	$status                = install_plugin_install_status( $data );
+	$requires_plugins      = isset( $data->requires_plugins ) ? $data->requires_plugins : array();
+	$plugin_dependency_met = true;
+
+	// Check if plugin dependency is installed and active.
+	foreach ( $requires_plugins as $dependency ) {
+		$plugin_dependency_met = false;
+		$active_plugins        = get_option( 'active_plugins' );
+		foreach ( $active_plugins as $plugin_file ) {
+			if ( str_contains( $plugin_file, '/' ) && explode( '/', $plugin_file )[0] === $dependency ) {
+				$plugin_dependency_met = true;
+				break;
+			}
+		}
+	}
 
 	sprintf(
 		'<a class="install-now button" data-slug="%s" href="%s" aria-label="%s" data-name="%s">%s</a>',
-		esc_attr( $data['slug'] ),
+		esc_attr( $data->slug ),
 		esc_url( $status['url'] ),
 		/* translators: %s: Plugin name and version. */
 		esc_attr( sprintf( _x( 'Install %s now', 'plugin' ), $name ) ),
@@ -38,10 +59,10 @@ function wp_get_plugin_action_button( $name, $data, $compatible_php, $compatible
 		switch ( $status['status'] ) {
 			case 'install':
 				if ( $status['url'] ) {
-					if ( $compatible_php && $compatible_wp ) {
+					if ( $compatible_php && $compatible_wp && $plugin_dependency_met ) {
 						$button = sprintf(
 							'<a class="install-now button" data-slug="%s" href="%s" aria-label="%s" data-name="%s">%s</a>',
-							esc_attr( $data['slug'] ),
+							esc_attr( $data->slug ),
 							esc_url( $status['url'] ),
 							/* translators: %s: Plugin name and version. */
 							esc_attr( sprintf( _x( 'Install %s now', 'plugin' ), $name ) ),
@@ -63,7 +84,7 @@ function wp_get_plugin_action_button( $name, $data, $compatible_php, $compatible
 						$button = sprintf(
 							'<a class="update-now button aria-button-if-js" data-plugin="%s" data-slug="%s" href="%s" aria-label="%s" data-name="%s">%s</a>',
 							esc_attr( $status['file'] ),
-							esc_attr( $data['slug'] ),
+							esc_attr( $data->slug ),
 							esc_url( $status['url'] ),
 							/* translators: %s: Plugin name and version. */
 							esc_attr( sprintf( _x( 'Update %s now', 'plugin' ), $name ) ),